requirements.txt
robotframework >= 3.2.1
pykafka >= 2.8.0
